requirements.txt
snowflake-connector-python==3.6.0
streamlit
pandas
python-dotenv
keyring